ai-line-bot/requirements.txt
flask==3.0.0
line-bot-sdk==3.7.0
python-dotenv==1.0.0
openai>=1.0.0
google-cloud-translate
google-cloud-speech
google-auth
google-auth-oauthlib